+++ netbox_config_diff/compliance/base.py
@@ -6,38 +6,43 @@ from typing import Iterable, Iterator
 from core.choices import DataSourceStatusChoices
 from core.models import DataFile, DataSource
 from dcim.choices import DeviceStatusChoices
-from dcim.models import Device, Site
-from django.core.exceptions import ObjectDoesNotExist
+from dcim.models import Device, DeviceRole, Site
+from django.conf import settings
 from django.db.models import Q
 from extras.scripts import MultiObjectVar, ObjectVar
 from jinja2.exceptions import TemplateError
 from netutils.config.compliance import diff_network_config
 from utilities.exceptions import AbortScript
 
-from netbox_config_diff.models import ConfigCompliance
+from netbox_config_diff.models import ConplianceDeviceDataClass
 
-from .models import DeviceDataClass
 from .secrets import SecretsMixin
-from .utils import PLATFORM_MAPPING, exclude_lines, get_unified_diff
+from .utils import PLATFORM_MAPPING, CustomChoiceVar, exclude_lines, get_unified_diff
 
 
 class ConfigDiffBase(SecretsMixin):
     site = ObjectVar(
         model=Site,
         required=False,
-        description="Run compliance for devices (with status Active, "
-        "primary IP, platform and config template) in this site",
+        description="Run compliance for devices (with primary IP, platform) in this site",
+    )
+    role = ObjectVar(
+        model=DeviceRole,
+        required=False,
+        description="Run compliance for devices with this role",
     )
     devices = MultiObjectVar(
         model=Device,
         required=False,
         query_params={
-            "status": DeviceStatusChoices.STATUS_ACTIVE,
             "has_primary_ip": True,
             "platform_id__n": "null",
-            "config_template_id__n": "null",
         },
-        description="If you define devices in this field, the Site field will be ignored",
+        description="If you define devices in this field, Site, Role fields will be ignored",
+    )
+    status = CustomChoiceVar(
+        choices=DeviceStatusChoices,
+        default=DeviceStatusChoices.STATUS_ACTIVE,
     )
     data_source = ObjectVar(
         model=DataSource,
@@ -56,8 +61,8 @@ class ConfigDiffBase(SecretsMixin):
         self.update_in_db(devices)
 
     def validate_data(self, data: dict) -> Iterable[Device]:
-        if not data["site"] and not data["devices"]:
-            raise AbortScript("Define site or devices")
+        if not data["site"] and not data["role"] and not data["devices"]:
+            raise AbortScript("Define site, role or devices")
         if data.get("data_source") and data["data_source"].status != DataSourceStatusChoices.COMPLETED:
             raise AbortScript("Define synced DataSource")
 
@@ -66,23 +71,29 @@ class ConfigDiffBase(SecretsMixin):
             devices = (
                 data["devices"]
                 .filter(
-                    status=DeviceStatusChoices.STATUS_ACTIVE,
+                    status=data["status"],
                     platform__platform_setting__isnull=False,
-                    config_template__isnull=False,
                 )
                 .exclude(
                     Q(primary_ip4__isnull=True) & Q(primary_ip6__isnull=True),
                 )
             )
         else:
-            devices = Device.objects.filter(
-                site=data["site"],
-                status=DeviceStatusChoices.STATUS_ACTIVE,
-                platform__platform_setting__isnull=False,
-                config_template__isnull=False,
-            ).exclude(
+            filters = {
+                "status": data["status"],
+                "platform__platform_setting__isnull": False,
+            }
+            if data["site"]:
+                filters["site"] = data["site"]
+            elif data["role"]:
+                if settings.VERSION.split(".", 1)[1].startswith("5"):
+                    filters["device_role"] = data["role"]
+                else:
+                    filters["role"] = data["role"]
+            devices = Device.objects.filter(**filters).exclude(
                 Q(primary_ip4__isnull=True) & Q(primary_ip6__isnull=True),
             )
+
         if data.get("devices"):
             if qs_diff := data["devices"].difference(devices):
                 platforms = {d.platform.name for d in qs_diff}
@@ -90,24 +101,18 @@ class ConfigDiffBase(SecretsMixin):
 
         if not devices:
             self.log_warning(
-                "No matching devices found, devices must have status `Active`, primary IP, platform and platformsetting"
+                "No matching devices found, devices must have status primary IP, platform and platformsetting"
             )
         else:
             self.log_info(f"Working with device(s): {', '.join(d.name for d in devices)}")
         return devices
 
-    def update_in_db(self, devices: list[DeviceDataClass]) -> None:
+    def update_in_db(self, devices: list[ConplianceDeviceDataClass]) -> None:
         for device in devices:
             self.log_results(device)
-            try:
-                obj = ConfigCompliance.objects.get(device_id=device.pk)
-                obj.snapshot()
-                obj.update(**device.to_db())
-                obj.save()
-            except ObjectDoesNotExist:
-                ConfigCompliance.objects.create(**device.to_db())
+            device.send_to_db()
 
-    def log_results(self, device: DeviceDataClass) -> None:
+    def log_results(self, device: ConplianceDeviceDataClass) -> None:
         if device.error:
             self.log_failure(f"{device.name} errored")
         elif device.diff:
@@ -115,11 +120,11 @@ class ConfigDiffBase(SecretsMixin):
         else:
             self.log_success(f"{device.name} no diff")
 
-    def get_devices_with_rendered_configs(self, devices: Iterable[Device]) -> Iterator[DeviceDataClass]:
+    def get_devices_with_rendered_configs(self, devices: Iterable[Device]) -> Iterator[ConplianceDeviceDataClass]:
         self.check_netbox_secrets()
         self.substitutes = {}
         for device in devices:
-            username, password = self.get_credentials(device)
+            username, password, auth_secondary = self.get_credentials(device)
             rendered_config = None
             error = None
             context_data = device.get_config_context()
@@ -138,7 +143,7 @@ class ConfigDiffBase(SecretsMixin):
                 if substitutes := device.platform.platform_setting.substitutes.all():
                     self.substitutes[platform] = [s.regexp for s in substitutes]
 
-            yield DeviceDataClass(
+            yield ConplianceDeviceDataClass(
                 pk=device.pk,
                 name=device.name,
                 mgmt_ip=str(device.primary_ip.address.ip),
@@ -147,11 +152,12 @@ class ConfigDiffBase(SecretsMixin):
                 exclude_regex=device.platform.platform_setting.exclude_regex,
                 username=username,
                 password=password,
+                auth_secondary=auth_secondary,
                 rendered_config=rendered_config,
                 error=error,
             )
 
-    def get_config_from_datasource(self, devices: list[DeviceDataClass]) -> None:
+    def get_config_from_datasource(self, devices: list[ConplianceDeviceDataClass]) -> None:
         for device in devices:
             if df := DataFile.objects.filter(source=self.data["data_source"], path__icontains=device.name).first():
                 if config := df.data_as_string:
@@ -161,14 +167,14 @@ class ConfigDiffBase(SecretsMixin):
             else:
                 device.error = f"Not found file in DataSource for device {device.name}"
 
-    def get_actual_configs(self, devices: list[DeviceDataClass]) -> None:
+    def get_actual_configs(self, devices: list[ConplianceDeviceDataClass]) -> None:
         if self.data["data_source"]:
             self.get_config_from_datasource(devices)
         else:
             loop = asyncio.get_event_loop()
             loop.run_until_complete(asyncio.gather(*(d.get_actual_config() for d in devices)))
 
-    def get_diff(self, devices: list[DeviceDataClass]) -> None:
+    def get_diff(self, devices: list[ConplianceDeviceDataClass]) -> None:
         for device in devices:
             if device.error is not None:
                 continue