+++ netbox_config_diff/models/data_models.py
@@ -5,40 +5,40 @@ from scrapli import AsyncScrapli
 
 from netbox_config_diff.choices import ConfigComplianceStatusChoices
 
+from .models import ConfigCompliance
+
 
 @dataclass
-class DeviceDataClass:
+class BaseDeviceDataClass:
     pk: int
     name: str
     mgmt_ip: str
     platform: str
     username: str
     password: str
-    command: str | None = None
     exclude_regex: str | None = None
     rendered_config: str | None = None
     actual_config: str | None = None
-    diff: str | None = None
+    diff: str = ""
     missing: str | None = None
     extra: str | None = None
-    error: str | None = None
+    error: str = ""
     config_error: str | None = None
     auth_strict_key: bool = False
+    auth_secondary: str | None = None
     transport: str = "asyncssh"
 
-    def __str__(self):
+    def __str__(self) -> str:
         return self.name
 
-    def __hash__(self):
-        return hash(self.name)
-
-    def to_scrapli(self):
+    def to_scrapli(self) -> dict:
         return {
             "host": self.mgmt_ip,
             "auth_username": self.username,
             "auth_password": self.password,
             "platform": self.platform,
             "auth_strict_key": self.auth_strict_key,
+            "auth_secondary": self.auth_secondary,
             "transport": self.transport,
             "transport_options": {
                 "asyncssh": {
@@ -78,13 +78,16 @@ class DeviceDataClass:
             },
         }
 
-    def to_db(self):
+    def get_status(self) -> str:
         if self.error:
-            status = ConfigComplianceStatusChoices.ERRORED
+            return ConfigComplianceStatusChoices.ERRORED
         elif self.diff:
-            status = ConfigComplianceStatusChoices.DIFF
+            return ConfigComplianceStatusChoices.DIFF
         else:
-            status = ConfigComplianceStatusChoices.COMPLIANT
+            return ConfigComplianceStatusChoices.COMPLIANT
+
+    def to_db(self) -> dict:
+        status = self.get_status()
 
         return {
             "device_id": self.pk,
@@ -97,7 +100,25 @@ class DeviceDataClass:
             "extra": self.extra or "",
         }
 
-    async def get_actual_config(self):
+    def send_to_db(self) -> None:
+        try:
+            obj = ConfigCompliance.objects.get(device_id=self.pk)
+            if obj.status != self.get_status():
+                obj.update(commit=True, **self.to_db())
+            elif obj.diff != self.diff or obj.error != self.error:
+                obj.update(commit=True, **self.to_db())
+        except ConfigCompliance.DoesNotExist:
+            ConfigCompliance.objects.create(**self.to_db())
+
+
+class ConplianceDeviceDataClass(BaseDeviceDataClass):
+    command: str
+
+    def __init__(self, command: str, **kwargs) -> None:
+        super().__init__(**kwargs)
+        self.command = command
+
+    async def get_actual_config(self) -> None:
         if self.error is not None:
             return
         try:
@@ -109,3 +130,8 @@ class DeviceDataClass:
                     self.actual_config = result.result
         except Exception:
             self.error = traceback.format_exc()
+
+
+class ConfiguratorDeviceDataClass(BaseDeviceDataClass):
+    def __hash__(self) -> int:
+        return hash(self.name)